cli: Display sensible artefact ranges
No point trying to be human friendly when that friendliness will bite
the user in the rear.  Make the start position zero-based, and the end
position one past the end:

  http://www.cs.utexas.edu/users/EWD/transcriptions/EWD08xx/EWD831.html

This way an artefact can be extracted from a file using dd:

  dd if=file of=extracted bs=1 skip=START count=$(( END - START ))

Note: this will only work if the input file does not contain
multi-byte characters (because dd wants byte positions, but we provide
character positions).