Don't automatically invoke cl:continue restart
These restarts are intended for interactive use, and should not be
invoked automatically.  For instance problems with opening files may
provide an interactive restart to try the file operation again with
intention that the user has fixed the problem with the file; invoking
this restart will not fix the problem with the file.